requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.19.0
werkzeug>=2.0.0